+++ src/main/java/caculator/string/StringSplitter.java
@@ -1,4 +1,4 @@
-package caculator;
+package caculator.string;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
