+++ src/main/java/caculator/string/StringSupport.java
@@ -1,4 +1,4 @@
-package caculator;
+package caculator.string;
 
 public class StringSupport {
