+++ src/test/java/calculator/string/StringSplitterTest.java
@@ -1,11 +1,11 @@
-package calculator;
+package calculator.string;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import caculator.StringSplitter;
+import caculator.string.StringSplitter;
 
 class StringSplitterTest {
