+++ src/test/java/calculator/string/StringTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.string;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import caculator.StringSupport;
+import caculator.string.StringSupport;
 
 class StringTest {
