+++ src/test/java/collection/SetTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package collection;
 
 import static org.assertj.core.api.Assertions.assertThat;
